+++ R/ZZZ.R
@@ -1,5 +1,11 @@
 .onAttach <- function(libname, pkgname) {
-  options(future.globals.maxSize = 100*1024^3)
+  options(future.globals = TRUE)
+  options(future.globals.maxSize = +Inf)
+  options(future.chunk.size = 1.0)
+  progressr::handlers(progressr::handler_progress(
+    format   = ":spin [:bar] :percent in :elapsed ETA: :eta",
+    complete = "="
+  ))
   data.table::setDTthreads(threads = 1, restore_after_fork = FALSE)
   packageStartupMessage("
 
@@ -12,6 +18,7 @@
 _/
 
                                 by: M\U00E1rton Kolossv\U00E1ry, MD PhD
+
 
 Please cite:
 Kolossv\U00E1ry M et al.