+++ R/convert_phy.R
@@ -42,11 +42,16 @@ convert_phy <- function(d, code = "phy_code", code_type = "phy_code_type",  code
   #Initialize multicore
   if(nThread == 1 | length(codes_to_find) == 1) {
     `%exec%` <- foreach::`%do%`
+    future::plan(future::sequential)
   } else {
     if(length(codes_to_find) > 0 & length(codes_to_find) < nThread) {nThread <- length(codes_to_find)}
-    cl <- parallel::makeCluster(nThread, methods = FALSE, useXDR = FALSE)
-    doParallel::registerDoParallel(cl)
-    `%exec%` <- foreach::`%dopar%`
+
+    if(parallelly::supportsMulticore()) {
+      future::plan(future::multicore, workers = nThread)
+    } else {
+      future::plan(future::multisession, workers = nThread)
+    }
+    `%exec%` <- doFuture::`%dofuture%`
   }
 
   #Create combined code colmn
@@ -59,7 +64,7 @@ convert_phy <- function(d, code = "phy_code", code_type = "phy_code_type",  code
 
   #Find diagnoses per row
   result <- foreach::foreach(i = 1:length(codes_to_find), .combine="cbind",
-                             .inorder=TRUE,
+                             .inorder=TRUE, .options.future = list(chunk.size = 1.0),
                              .errorhandling = c("pass"), .verbose=FALSE) %exec%
     {
       if(is.null(collapse)) {
@@ -84,7 +89,7 @@ convert_phy <- function(d, code = "phy_code", code_type = "phy_code_type",  code
         diag_coll
       }
     }
-  if(exists("cl") & nThread>1) {parallel::stopCluster(cl)}
+  future::plan(future::sequential)
 
   if(is.null(collapse)) { #Remove unnecessary info and combine with original data if non-collapse
     result <- cbind(d, result)