+++ R/convert_rfv.R
@@ -37,11 +37,16 @@ convert_rfv <- function(d, code = "rfv_concept_id", codes_to_find = NULL,
   #Initialize multicore
   if(nThread == 1 | length(codes_to_find) == 1) {
     `%exec%` <- foreach::`%do%`
+    future::plan(future::sequential)
   } else {
     if(length(codes_to_find) > 0 & length(codes_to_find) < nThread) {nThread <- length(codes_to_find)}
-    cl <- parallel::makeCluster(nThread, methods = FALSE, useXDR = FALSE)
-    doParallel::registerDoParallel(cl)
-    `%exec%` <- foreach::`%dopar%`
+
+    if(parallelly::supportsMulticore()) {
+      future::plan(future::multicore, workers = nThread)
+    } else {
+      future::plan(future::multisession, workers = nThread)
+    }
+    `%exec%` <- doFuture::`%dofuture%`
   }
 
   #Find diagnoses if requested
@@ -50,7 +55,7 @@ convert_rfv <- function(d, code = "rfv_concept_id", codes_to_find = NULL,
 
     #Find diagnoses per row
     result <- foreach::foreach(i = 1:length(codes_to_find), .combine="cbind",
-                               .inorder=TRUE,
+                               .inorder=TRUE, .options.future = list(chunk.size = 1.0),
                                .errorhandling = c("pass"), .verbose=FALSE) %exec%
       {
         if(is.null(collapse)) {
@@ -75,7 +80,7 @@ convert_rfv <- function(d, code = "rfv_concept_id", codes_to_find = NULL,
           diag_coll
         }
       }
-    if(exists("cl") & nThread>1) {parallel::stopCluster(cl)}
+    future::plan(future::sequential)
 
     if(is.null(collapse)) { #Remove unnecessary info and combine with original data if non-collapse
       result <- cbind(d, result)
@@ -85,7 +90,7 @@ convert_rfv <- function(d, code = "rfv_concept_id", codes_to_find = NULL,
     }
     return(result)
   } else { #If no diagnoses
-    if(exists("cl") & nThread>1) {parallel::stopCluster(cl)}
+    future::plan(future::sequential)
     return(d)
   }
 }