+++ tests/covr.R
@@ -1,5 +1,5 @@
 # library(covr)
-# coverage <- covr::package_coverage(path="/Users/mjk2/Dropbox (Personal)/R_packages/parseRPDR/", quiet = FALSE)
+# coverage <- covr::package_coverage(path="/Users/mjk2/Dropbox (Personal)/R_packages/parseRPDR/")
 # coverage
 # report(coverage)
 # covr::codecov(token = "YOUR_TOKEN_HERE")