+++ tests/testthat.R
@@ -1,4 +1,4 @@
 library(testthat)
 library(parseRPDR)
 
-testthat::test_check("parseRPDR")
+# testthat::test_check("parseRPDR")