+++ tests/testthat/test-convert_enc.R
@@ -6,7 +6,7 @@
 ## Date: 2023-02-24
 ###############################################################################
 
-# testthat::skip_if_offline()
+testthat::skip_if_offline()
 # Load and check equality ====================
 
 suppressPackageStartupMessages(library(parseRPDR))
@@ -43,9 +43,17 @@ test_that("convert_enc run using sequential and parallel loading returns same re
 })
 
 expect_true({
-  c_p_sum <- suppressMessages(convert_enc(d_p, codes_to_find = diseases, collapse = "ID_MERGE", aggr_type = "latest", nThread = 2))
-  c_p_sum <- suppressMessages(convert_enc(d_p, codes_to_find = NULL, collapse = "ID_MERGE", aggr_type = "latest", nThread = 2))
+  c_p_sum <- suppressMessages(convert_enc(d_p, codes_to_find = diseases, collapse = "ID_MERGE", aggr_type = "earliest", nThread = 2))
+  c_p_sum <- suppressMessages(convert_enc(d_p, codes_to_find = NULL, collapse = "ID_MERGE", aggr_type = "earliest", nThread = 2))
   TRUE
+})
+
+c_s_sum <- suppressMessages(convert_enc(d_p, codes_to_find = diseases, collapse = "ID_MERGE", aggr_type = "latest", nThread = 1))
+c_p_sum <- suppressMessages(convert_enc(d_p, codes_to_find = diseases, collapse = "ID_MERGE", aggr_type = "latest", nThread = 2))
+
+
+test_that("convert_enc summarizing using sequential and parallel loading returns same results", {
+  expect_equal(c_s_sum, c_p_sum)
 })
 
 c_s_sum <- suppressMessages(convert_enc(d_p, codes_to_find = diseases, collapse = "ID_MERGE", aggr_type = "earliest", nThread = 1))